src/backend/data/requirements.txt
kafka-python
pymongo
PyMuPDF
requests
tqdm
python-dotenv
typing_extensions
